Fix some bad casting in error reporting
For larger-architectures, it's probably more advisable to use a long.
The types were also wrong, because casting has a higher precedence
order than the infix operator '+'.

The format string %Zu is the cleanest solution, but not defined in
many compilers and in C89, so compromise using this well-known cast
heuristic that works 'pretty well' on 'most platforms' and 'most
compilers'.

Signed-off-by: Daniel Farina <daniel@heroku.com>