xlogparse: parameterize page info output
As a library, our parsing code shouldn't assume that clients want
stdout descriptions of each page as it's read.  To fix this and
support the traditional xlogdump model, this behavior can now be
optionally enabled using the SetPageInfoDump() routine.  In the
future, it may make sense to take this a step further and give
clients a finer grained control, like the ability to specify
the output stream for both standard and error output.